requirements.txt
streamlit
PyPDF2
pdf2image
pytesseract
pillow
numpy<2